+++ src/ui/action_handler.py
@@ -1,5 +1,6 @@
 from PyQt6.QtWidgets import QDialog, QFileDialog
 
+from operations.heic2jpg import heic_to_jpg_operation
 from operations.takeout import takeout_operation
 from ui.browser import Browser
 from ui.folder_select import FolderSelectWidget
@@ -15,6 +16,7 @@ class ActionHandler:
 
     def setup_slots(self):
         self.action_buttons['takeout'].pressed.connect(self.handle_takeout)
+        self.action_buttons['heic2jpg'].released.connect(self.handle_heic2jpg)
         # self.btn_number.pressed.connect(self.numberButtonAction)
         # self.btn_rename.pressed.connect(self.renameButtonAction)
         # self.btn_resize.pressed.connect(self.resizeButtonAction)
@@ -70,6 +72,48 @@ class ActionHandler:
         # perform follow-up actions after closing
         if dlg.go_to_output:
             self.folder_select.force_set_directory(dlg.edit_output_path.text())
+
+    def handle_heic2jpg(self):
+        from ui.designer.heic2jpg import Ui_heic2jpg
+
+        # define a wrapper class for the dialog
+        class Heic2JpegDialog(QDialog, Ui_heic2jpg):
+            def __init__(self, initial_folder: str, parent=None):
+                QDialog.__init__(self, parent)
+                self.setupUi(self)
+                self.edit_folder_path.setText(initial_folder)
+                self.go_to_output = False
+
+                # slots
+                self.btn_perform_action.clicked.connect(self.start_heic2jpg)
+                self.btn_folder_select.clicked.connect(
+                    lambda: self.edit_folder_path.setText(
+                        QFileDialog.getExistingDirectory(self, 'Select Images Folder', self.edit_folder_path.text())
+                    )
+                )
+                self.btn_close_redirect.clicked.connect(self.on_close_redirect)
+
+            def start_heic2jpg(self):
+                def callback(message, progress):
+                    self.progress_bar.setValue(progress)
+                    self.text_output.append(message)
+
+                # fetch the input and clean the dialog log textedit
+                folder_path = self.edit_folder_path.text()
+                self.text_output.clear()
+                # call the takeout operation
+                return heic_to_jpg_operation(folder_path, callback)
+
+            def on_close_redirect(self):
+                self.go_to_output = True
+                self.accept()
+
+        # create the dialog and show it,
+        dlg = Heic2JpegDialog(self.folder_select.folder_edit.text(), self.parent)
+        dlg.exec()
+        # perform follow-up actions after closing
+        if dlg.go_to_output:
+            self.folder_select.force_set_directory(dlg.edit_folder_path.text())
 
     # def handle_import(self):
     #     files = self.browser.get_selection()